dandelion-tower: improve API (#257)
* init

* reduce the jobs handled by the dandelion pool

* fix docs

* resolve todo

* review changes

* Update p2p/dandelion-tower/src/pool/incoming_tx.rs


* Update p2p/dandelion-tower/src/pool/incoming_tx.rs


* `PId` -> `PeerId`

---------
